refactor: 노래 삭제 버튼 위치 변경
- 노래 리스트에서 삭제할 수 있도록 변경
